Splitted profile into multiple components.
