+++ FrontEnd/src/Components/Profile/CustomProfileSnackbars.jsx
@@ -1,0 +1,35 @@
+import React from 'react';
+import MuiAlert from '@mui/material/Alert';
+import Snackbar from '@mui/material/Snackbar';
+
+const Alert = React.forwardRef((props, ref) => <MuiAlert elevation={6} ref={ref} variant="filled" {...props} />);
+
+function CustomProfileSnackbars({
+  handleClose, successOpen, errorOpen, infoOpen, alertMessage,
+}) {
+  return (
+    <>
+      <Snackbar
+        open={successOpen}
+        autoHideDuration={4000}
+        onClose={handleClose}
+      >
+        <Alert onClose={handleClose} severity="success" sx={{ width: '100%' }}>
+          {alertMessage}
+        </Alert>
+      </Snackbar>
+      <Snackbar open={errorOpen} autoHideDuration={4000} onClose={handleClose}>
+        <Alert onClose={handleClose} severity="error" sx={{ width: '100%' }}>
+          {alertMessage}
+        </Alert>
+      </Snackbar>
+      <Snackbar open={infoOpen} autoHideDuration={4000} onClose={handleClose}>
+        <Alert onClose={handleClose} severity="info" sx={{ width: '100%' }}>
+          {alertMessage}
+        </Alert>
+      </Snackbar>
+    </>
+  );
+}
+
+export default CustomProfileSnackbars;
